feat(site): design error page
