+++ src/site/pages/posts/+guard.ts
@@ -1,10 +1,10 @@
 import { render } from 'vike/abort';
 import type { GuardAsync } from 'vike/types';
-import { postsRouteInfos } from '../../routes/posts-route.js';
+import type { PostsRouteParams } from '../../routes/posts-route.js';
+import { postsRoute } from '../../routes/posts-route.js';
 
 export const guard: GuardAsync = async (pageContext): ReturnType<GuardAsync> => {
-  const { managerName } = pageContext.routeParams;
-  if (!managerName || !Object.keys(postsRouteInfos).includes(managerName)) {
+  if (!postsRoute.guard?.(pageContext.routeParams as PostsRouteParams)) {
     throw render(404);
   }
 };
