+++ src/site/routes/error-route.ts
@@ -1,0 +1,9 @@
+import type { SiteRoute } from '../../core/entities/site-route.js';
+
+export const errorRoute: SiteRoute<never> = {
+  path: '/*',
+  info: () => ({
+    title: 'Error',
+  }),
+  createUrl: () => '/error/',
+};
