+++ src/site/routes/index.ts
@@ -1,4 +1,5 @@
 import { resolveRoute } from 'vike/routing';
+import { errorRoute } from './error-route.js';
 import { helpRoute } from './help-route.js';
 import { homeRoute } from './home-route.js';
 import { postRoute } from './post-route.js';
@@ -6,11 +7,13 @@ import { postsRoute } from './posts-route.js';
 import { userRoute } from './user-route.js';
 import { usersRoute } from './users-route.js';
 
-export const routes = [homeRoute, helpRoute, usersRoute, userRoute, postRoute, postsRoute];
+export const routes = [homeRoute, helpRoute, usersRoute, userRoute, postRoute, postsRoute, errorRoute];
 
 export function resolveFirstRoute(pathname: string) {
   for (const route of routes) {
     const { match, routeParams } = resolveRoute(route.path, pathname);
+    if (route.guard && !route.guard(routeParams as never)) continue;
+
     if (match) {
       return { route, params: route.mapParams ? route.mapParams(routeParams) : routeParams };
     }
